рефакторинг 06.02.20 9.32
